backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
tensorflow>=2.15.0
torch>=2.0.1
torchvision>=0.15.0
transformers>=4.33.2
Pillow>=9.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
pandas>=2.0.3
numpy>=1.24.3
Pillow>=10.0.0
opencv-python>=4.8.1.78
requests>=2.31.0
tweepy>=4.14.0
python-dotenv>=1.0.0
geopy>=2.3.0
folium>=0.14.0
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.16.1
nltk>=3.8.1
textblob>=0.17.1
spacy>=3.6.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
pytest>=7.4.2
pytest-flask>=1.2.0
gunicorn>=21.2.0 